backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
python-dotenv
pydantic-settings
qrcode
python-multipart